fix to ids

+++ utils/callbacks.py
@@ -188,7 +188,7 @@ class EvalCallback():
                 #------------------------------#
                 #   获得预测框
                 #------------------------------#
-                gt_boxes    = np.array([np.array(list(map(float,box.split(',')))) for box in line[1:]])
+                gt_boxes    = np.array([np.array(list(map(int,box.split(',')))) for box in line[1:]])
                 #------------------------------#
                 #   获得预测txt
                 #------------------------------#
